requirements.txt
google-cloud-firestore
flask-cors
gunicorn
flask
